multicoresplitkmatmul

+++ examples/102_dynamic_optimized_matmul/include/tiling_params.h
@@ -93,7 +93,7 @@ struct TilingParams {
     uint16_t k1{0};
     uint8_t swizzleOffset{1};
     uint8_t swizzleDirection{0};
-    uint16_t splitkFactor{1};
+    uint8_t splitkFactor{1};
     // The following parameters are only used in tiling and are not read by the kernel.
     uint8_t layoutTagA;
     uint8_t layoutTagB;
